Benchmarks: add benchmarks for commit 0bf8
As expected, the change slows down the solver because subspace
iterations are more expensive and because the search space grows, e.g.,
for the matrix pencil gyro_k/gyro_m, the final search space size
increased from 2354 to 3444 (+46%) and for bmwcra_1 from 116 to 198
(+71%). Consequently, the solver ran out of memory four times but
surprisingly not on the test problem pwtk with 218,000 DOF.

The maximum relative forward error is often significantly decreased and
below 1 with the exception of vanbody (2.13 * 10**4).