Licensing: remove Exhibit B (MPL 2.0)
What I wanted to achieve by enacting Exhibit B of the Mozilla Public
License 2.0 is to prevent someone from just re-licensing the code under
a GNU license. What I did achieve was GNU license incompatibility.
The MPL 2.0 does not allow such a blatant re-licensing attempt
(I guess this is a consequence of Article 3.4) and this is explicitly
said in the MPL 2.0 FAQ, Q14, paragraph 2:
  https://www.mozilla.org/en-US/MPL/2.0/FAQ/#mpl-and-lgpl